requirements.txt
PyPDF2~=3.0.1
pillow~=11.1.0
pdf2image~=1.17.0
pdfrw~=0.4
reportlab~=4.3.1
